#117 - Workout entry should be supported

+++ irunninglog-strava/src/main/java/com/irunninglog/strava/IStravaAthlete.java
@@ -2,9 +2,9 @@ package com.irunninglog.strava;
 
 public interface IStravaAthlete {
 
-    long getId();
+    int getId();
 
-    IStravaAthlete setId(long id);
+    IStravaAthlete setId(int id);
 
     String getEmail();
